All postman tests except stretch goals pass!

+++ Keepr/Services/VaultKeepsService.cs
@@ -11,15 +11,18 @@ namespace Keepr.Services
             _vaultsService = vaultsService;
         }
 
-        internal VaultKeep Create(VaultKeep vaultKeepData)
+        internal VaultKeep Create(VaultKeep vaultKeepData, string userId)
         {
             VaultKeep vaultKeep = _repo.Create(vaultKeepData);
+            Vault vault = _vaultsService.GetOne(vaultKeep.VaultId, userId);
+            if(vault.CreatorId != userId) throw new Exception("You cannot do this to me");
             return vaultKeep;
         }
 
         internal List<VaultKeepViewModel> GetKeeps(int vaultId)
         {
             List<VaultKeepViewModel> keeps = _repo.GetKeepsInVault(vaultId);
+            if(keeps == null) throw new Exception("Keep does not exist");
             return keeps;
         }
 
@@ -28,9 +31,9 @@ namespace Keepr.Services
             VaultKeep vaultKeep = _repo.GetOne(id);
             if(vaultKeep == null) throw new Exception("No VaultKeep found");
             if(vaultKeep.CreatorId != userId) throw new Exception("Not your VaultKeep");
-            _repo.Remove(id);
+            if(vaultKeep != null) _repo.Remove(id);
+            Console.WriteLine(vaultKeep);
             return "VaultKeep removed";
-
         }
   }
 }
